Cell Type-Specific Spatial Variability Analysis

Performs cell type-specific spatial gene variability analysis

Usage

spacelink_cell_type(
  normalized_counts,
  spatial_coords,
  cell_type_proportions,
  focal_cell_type,
  covariates = NULL,
  global_spacelink_results = NULL,
  lengthscales = NULL,
  n_lengthscales = 5,
  M = 1,
  calculate_ESV = TRUE,
  c1 = 0.25,
  c2 = 0.2,
  n_workers = 1
)

Arguments

normalized_counts

Normalized count matrix. Rows and columns indicate genes and spots, respectively.

spatial_coords

Spatial coordinates matrix. Rows indicate spots.

cell_type_proportions

Cell type proportion matrix (e.g., estimated by RCTD).

focal_cell_type

Name of cell-type to be tested.

covariates

Spatial features (without intercept)

global_spacelink_results

Spacelink global SVG test results from spacelink_global.

lengthscales

If NULL, length-scales are calculated by using our two-step approach.

n_lengthscales

Number of length-scales (i.e., number of kernels)

M

It controls the minimum length-scale. Minimum length-scale is set to be the minimum distance multiplied by M.

calculate_ESV

If TRUE, ct-ESV is returned.

c1

Gating parameter 1. Default is 0.25.

c2

Gating parameter 2. Default is 0.2.

n_workers

Number of workers for parallelization. Default is 1.

Value

A data frame containing:

time

Computation time

pval

Combined p-value for cell type-specific spatial patterns

ESV

ct-ESV for cell type-specific spatial patterns

Examples

data(Visium_human_DLPFC)
counts <- Visium_human_DLPFC$counts
spatial_coords <- Visium_human_DLPFC$spatial_coords
# Filter mitochondrial and low-expressed genes
counts <- counts[!grepl("(^MT-)|(^mt-)", rownames(counts)),]
counts <- counts[rowSums(counts >= 3) >= ncol(counts)*0.005,]
#> Error in h(simpleError(msg, call)): error in evaluating the argument 'i' in selecting a method for function '[': 'x' must be an array of at least two dimensions
# Normalize expression counts using sctransform package
library(Seurat)
#> Error in library(Seurat): there is no package called ‘Seurat’
library(sctransform)
#> Error in library(sctransform): there is no package called ‘sctransform’
seurat_obj <- CreateSeuratObject(counts = counts)
#> Error in CreateSeuratObject(counts = counts): could not find function "CreateSeuratObject"
seurat_norm = SCTransform(seurat_obj, vst.flavor = "v2", verbose = FALSE)
#> Error in SCTransform(seurat_obj, vst.flavor = "v2", verbose = FALSE): could not find function "SCTransform"
counts <- seurat_norm@assays$SCT$data
#> Error: object 'seurat_norm' not found
cell_type_proportions <- Visium_human_DLPFC$cell_type_proportions
cell_type_spacelink_results <- spacelink_cell_type(
  normalized_counts = counts[1:5,],
  spatial_coords = spatial_coords,
  cell_type_proportions = cell_type_proportions,
  focal_cell_type = "oligodendrocyte",
  calculate_ESV = TRUE
)
#> Error in rowMeans(Y): 'x' must be an array of at least two dimensions