Global Spatial Variability Analysis

Performs global spatial gene variability analysis across multiple length scales using kernel-based methods.

Usage

spacelink_global(
  normalized_counts,
  spatial_coords,
  covariates = NULL,
  lengthscales = NULL,
  n_lengthscales = 5,
  M = 1,
  n_workers = 1
)

Arguments

normalized_counts

Normalized count matrix. Rows and columns indicate genes and spots, respectively.

spatial_coords

Spatial coordinates matrix. Rows indicate spots.

covariates

Spatial features (without intercept). Rows indicate spots.

lengthscales

If NULL, length-scales are calculated using the two-step approach.

n_lengthscales

Number of length-scales (i.e., number of kernels). Default is 5.

M

Controls the minimum length-scale. Minimum length-scale is set to be the minimum distance multiplied by M. Default is 1.

n_workers

Number of workers for parallelization. Default is 1.

Value

A data frame containing:

tau.sq

Nugget variance component

sigma.sq1, sigma.sq2, ...

Spatial variance components for each kernel

raw_ESV

Effective Spatial Variability score

pval

Combined p-value from score tests

padj

Benjamini-Hochberg adjusted p-values

ESV

ESV adjusted for multiple testing (0 if padj > 0.05)

Examples

data(Visium_human_DLPFC)
counts <- Visium_human_DLPFC$counts
spatial_coords <- Visium_human_DLPFC$spatial_coords
# Filter mitochondrial and low-expressed genes
counts <- counts[!grepl("(^MT-)|(^mt-)", rownames(counts)),]
counts <- counts[rowSums(counts >= 3) >= ncol(counts)*0.005,]
#> Error in h(simpleError(msg, call)): error in evaluating the argument 'i' in selecting a method for function '[': 'x' must be an array of at least two dimensions
# Normalize expression counts using sctransform package
library(Seurat)
#> Error in library(Seurat): there is no package called ‘Seurat’
library(sctransform)
#> Error in library(sctransform): there is no package called ‘sctransform’
seurat_obj <- CreateSeuratObject(counts = counts)
#> Error in CreateSeuratObject(counts = counts): could not find function "CreateSeuratObject"
seurat_norm = SCTransform(seurat_obj, vst.flavor = "v2", verbose = FALSE)
#> Error in SCTransform(seurat_obj, vst.flavor = "v2", verbose = FALSE): could not find function "SCTransform"
counts <- seurat_norm@assays$SCT$data
#> Error: object 'seurat_norm' not found
global_results <- spacelink_global(normalized_counts = counts[1:5,], spatial_coords = spatial_coords)
#> Error in rowMeans(Y): 'x' must be an array of at least two dimensions